requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Assets==2.1.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-Babel==3.1.0
Werkzeug==3.0.1
itsdangerous==2.1.2
email-validator==2.1.0.post1
PyJWT==2.8.0
passlib==1.7.4

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pymysql==1.1.0  # For MySQL

# Web Scraping & API Clients
requests==2.32.4
beautifulsoup4==4.13.4
lxml==6.0.0
selenium==4.34.2
webdriver-manager==4.0.2
Scrapy==2.13.3
python-amazon-paapi==5.0.1
ebay_rest==1.0.13

# Data Processing & Utilities
pandas==2.3.1
numpy==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
python-slugify==8.0.1
bleach==6.1.0
markdown==3.5.1

# Background Tasks
celery==5.3.6
redis==5.0.1  # For Celery broker
flower==2.0.1  # For monitoring Celery

# Authentication & Security
cryptography==45.0.5
pyOpenSSL==25.1.0
python-dotenv==1.1.1
PySocks==1.7.1  # For proxy support

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
Faker==19.13.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinx-autodoc-typehints==1.25.1

# Deployment
gunicorn==21.2.0
gevent==23.7.0
supervisor==4.2.5

# Additional Dependencies
Pillow==10.1.0
boto3==1.28.65  # For AWS S3 storage
sendgrid==6.11.0  # For email sending
twilio==8.9.0  # For SMS notifications

# Existing Dependencies
attrs==25.3.0
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.34.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.0
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
loguru==0.7.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
queuelib==1.8.0
regex==2025.7.34
requests-file==2.1.0
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.16.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
tldextract==5.3.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
trio==0.30.0
trio-websocket==0.12.2
triton==3.3.1
Twisted==25.5.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
w3lib==2.3.1
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
